test/requirements.txt
pytest==8.1.1
cocotb==1.8.1
Pillow==10.4.0
